Add files via upload
Hi everyone, 

I picked interest in knowing the annual cost of keeping inmates who commit crimes. This is a big problem many governments face. The judiciary also has a backlog of cases that are reported daily which may cause delays and inconveniences. check this new york times article: https://www.nytimes.com/2013/08/24/nyregion/citys-annual-cost-per-inmate-is-nearly-168000-study-says.html
(City’s Annual Cost Per Inmate Is $168,000, Study Finds). They reported that the annual average taxpayer cost in these states was $31,286 per inmate. This quite a lot of money if you have many inmates. 

In our data set, I selected wrote a program to count every type of crime if and only if the arrest was made (In this case True bool), the results can be found in the crimes.txt file. I decided to use a benchmark value of $31,286 per crime committed. The cost is added in the distributed cost.txt cache file.  Assuming that all the arrested criminals were to spend at least one year in prison, how much would the government spend on every crime in total? The answer to this question helps the officials to know how much they have spent on every crime since 2001.
I hadoop mappers and reducers as named in the multiple input classes. You will see that the code uses a distributed cache which is very resourceful when you don't want to distribute the cache file onto multiple nodes. The results can be found in the part-r-00000 file. 
Thank you.